Order user comments by updated_at.
We need to set order for comments collection to make sure we constantly
getting comments in the same order.

> If sorting is not chosen, the rows will be returned in an unspecified order.
The actual order in that case will depend on the scan and join plan types and
the order on disk, but it must not be relied on. A particular output
ordering can only be guaranteed if the sort step is explicitly chosen.

http://www.postgresql.org/docs/9.4/static/queries-order.html